+++ R/globalVariables.R
@@ -9,11 +9,12 @@ globalVariables(
     "proust_char", 
     "books", 
     "sentence", 
-    "sentiments_polarity", 
-    "sentiment_score", 
     "stop_words_fr", 
     "full_list", 
     "vol", 
     "count", 
+    "text", 
+    ".", 
+    "vec", 
     "text")
 )