+++ R/proust_sentiments.R
@@ -1,23 +1,14 @@
-#' Sentiment Lexicon
+#' Old sentiment lexicon
+#
+#' This function has been deprecated, and will be in next proustr version.
+#' See the {rfeel} package now: http://github.com/ColinFay/rfeel
 #' 
-#' A sentiment lexicon with eitheir polarity or score.
-#'
-#' @param type polarity (positive or negative) or score on six sentiments (joy, fear, sadness, anger, surprise, disgust)
-#' 
-#' @source Amine Abdaoui, Jérôme Azé, Sandra Bringay et Pascal Poncelet. FEEL: French Expanded Emotion Lexicon. Language Resources and Evaluation, LRE 2016, pp 1-23.
-#'
-#' @return a tibble
+#' @param type For backward compatibility
 #' @export
-#'
-#' @examples
-#' proust_sentiments(type = "score")
-#' proust_sentiments(type = "polarity")
+#' @return a tibble
 
 proust_sentiments <- function(type = c("polarity", "score")) {
-  type <- match.arg(type)
-  if (type == "polarity"){
-    structure(proustr::sentiments_polarity, class = c("tbl_df", "tbl", "data.frame"))
-  } else {
-    structure(proustr::sentiments_score, class = c("tbl_df", "tbl", "data.frame"))
-  }
+  message("This function no longer lives in {proustr}.")
+  message("See the {rfeel} package.")
+  message("http://github.com/ColinFay/rfeel")
 }